requirements.txt
aiofiles==0.6.0
aiohttp==3.7.3
alembic==1.5.4
amqp==5.0.5
appdirs==1.4.4
APScheduler==3.7.0
async-timeout==3.0.1
attrs==20.3.0
autoflake==1.4
autopep8==1.5.4
beautifulsoup4==4.9.3
billiard==3.6.3.0
black==20.8b1
bs4==0.0.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
fastapi==0.63.0
fastapi-utils==0.2.1
flake8==3.8.4
h11==0.12.0
idna==2.10
isort==5.7.0
Jinja2==2.11.3
kombu==5.0.2
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==5.1.0
mypy==0.800
mypy-extensions==0.4.3
orjson==3.4.7
packaging==20.9
pathspec==0.8.1
Pillow==8.1.0
pkg-resources==0.0.0
pony==0.7.14
prompt-toolkit==3.0.14
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pydantic==1.7.3
pyflakes==2.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-multipart==0.0.5
pytz==2021.1
qrcode==6.1
regex==2020.11.13
requests==2.25.1
six==1.15.0
soupsieve==2.2
SQLAlchemy==1.3.22
sqlalchemy-stubs==0.4
starlette==0.13.6
stripe==2.55.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.2
uvicorn==0.13.3
vine==5.0.0
wcwidth==0.2.5
yarl==1.6.3
